requirements.txt
alembic==1.15.1
aniso8601==10.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
flask-sock==0.7.0
Flask-SQLAlchemy==3.1.1
Flask-Turbo==0.0.0
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.2.0
oauthlib==3.2.2
packaging==24.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.38
Turbo-Flask==0.8.6
typing_extensions==4.12.2
urllib3==2.3.0
URLObject==2.4.3
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
